Changed scores table and the way that quiz scores are validated
The columns of scores table were changed to store questions by a question id and a boolean value to say if that user answered the question correctly.
Previously, the quiz make sure all of the user's answers were correct before redirecting. Now the quiz allows users to enter wrong answers and redirects to the results page.